backend/requirements.txt
pandas
SQLAlchemy>=2.0
psycopg2-binary
python-dotenv

google-generativeai
